Fix dataloader
