+++ test/dataloader/dataloader.py
@@ -2,13 +2,13 @@ import os; os.environ['XLA_FLAGS'] = os.environ.get('XLA_FLAGS', '') + ' --xla_f
 import jax; jax.config.update('jax_platforms', 'cpu')
 from pathlib import Path; import sys; sys.path.append(str(Path(__file__).resolve().parent.parent.parent))
 
-from lib.dataloader.dataloader import dataloader
+from lib.dataloader.DataLoader import DataLoader
 from lib.random.wrapper import seed2key
 
 if __name__ == '__main__':
     key = seed2key(42)
 
-    data_loader = dataloader(dataset='dummy', batch_size=48, key=key, n_workers=32)
+    data_loader = DataLoader(dataset='dummy', key=key, batch_size=48, n_workers=32)
     for n_batches, batch in enumerate(data_loader):
         print(
             batch.src.shape,
